+++ app/Http/Controllers/api/SestraController.php
@@ -6,9 +6,10 @@ use App\Models\User;
 use App\Models\Doktor;
 use App\Models\Sestra;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
-use App\Http\Resources\Resource\SestraResource;
 use App\Trait\CanLoadRelationships;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Resources\Resource\SestraResource;
 
 class SestraController extends Controller
 {
@@ -20,6 +21,7 @@ class SestraController extends Controller
     public function index()
     {
         $sestre = Sestra::query();
+        Gate::authorize('viewAny', Sestra::class);
         return SestraResource::collection($sestre->latest()->paginate());
     }
 
@@ -29,32 +31,34 @@ class SestraController extends Controller
     public function store(Request $request)
     {
         // Validacija za User
-    $validatedUser = $request->validate([
-        'name' => 'required|string|max:20',
-        'email' => 'required|email|unique:users',
-        'password' => 'required|string|min:8',
-    ]);
+        Gate::authorize('create', Sestra::class);
 
-    // Kreiranje User-a
-    $user = User::create([
-        'name' => $validatedUser['name'],
-        'email' => $validatedUser['email'],
-        'password' => bcrypt($validatedUser['password']), // Šifriranje lozinke
-        'role' => 'sestra', // Postavljanje role
-    ]);
+        $validatedUser = $request->validate([
+            'name' => 'required|string|max:20',
+            'email' => 'required|email|unique:users',
+            'password' => 'required|string|min:8',
+        ]);
+
+        // Kreiranje User-a
+        $user = User::create([
+            'name' => $validatedUser['name'],
+            'email' => $validatedUser['email'],
+            'password' => bcrypt($validatedUser['password']), // Šifriranje lozinke
+            'role' => 'sestra', // Postavljanje role
+        ]);
 
 
 
-    $sestra = Sestra::create([
-        'user_id' => $user->id
-    ]);
+        $sestra = Sestra::create([
+            'user_id' => $user->id
+        ]);
 
-    // Povratni odgovor
-    return response()->json([
-        'message' => 'Sestra je uspešno kreirana.',
-        'user' => $user,
-        'sestra' => $sestra,
-    ], 201);
+        // Povratni odgovor
+        return response()->json([
+            'message' => 'Sestra je uspešno kreirana.',
+            'user' => $user,
+            'sestra' => $sestra,
+        ], 201);
     }
 
     /**
@@ -63,6 +67,7 @@ class SestraController extends Controller
     public function show(string $id)
     {
         $sestra = Sestra::findOrFail($id);
+        Gate::authorize('view', $sestra);
         return new SestraResource($sestra);
     }
 
@@ -72,30 +77,32 @@ class SestraController extends Controller
     public function update(Request $request, string $id)
     {
         // Pronađi doktora i povezanog korisnika
-    $sestra = Sestra::findOrFail($id);
-    $user = User::findOrFail($sestra->user_id);
-
-    // Validacija za User-a
-    $validatedUser = $request->validate([
-        'name' => 'required|string|max:20',
-        'email' => 'required|email|unique:users,email,' . $user->id, // Ignoriši trenutni korisnik
-        'password' => 'nullable|string|min:8', // Lozinka nije obavezna prilikom ažuriranja
-    ]);
-
-    // Ažuriranje User-a
-    $user->update([
-        'name' => $validatedUser['name'],
-        'email' => $validatedUser['email'],
-        'password' => $validatedUser['password'] ? bcrypt($validatedUser['password']) : $user->password, // Ako nema nove lozinke, zadrži staru
-    ]);
+        $sestra = Sestra::findOrFail($id);
+        $user = User::findOrFail($sestra->user_id);
+        Gate::authorize('update', $sestra);
 
 
-    // Povratni odgovor
-    return response()->json([
-        'message' => 'Sestra je uspešno ažurirana.',
-        'user' => $user,
-        'sestra' => $sestra,
-    ], 200);
+        // Validacija za User-a
+        $validatedUser = $request->validate([
+            'name' => 'required|string|max:20',
+            'email' => 'required|email|unique:users,email,' . $user->id, // Ignoriši trenutni korisnik
+            'password' => 'nullable|string|min:8', // Lozinka nije obavezna prilikom ažuriranja
+        ]);
+
+        // Ažuriranje User-a
+        $user->update([
+            'name' => $validatedUser['name'],
+            'email' => $validatedUser['email'],
+            'password' => $validatedUser['password'] ? bcrypt($validatedUser['password']) : $user->password, // Ako nema nove lozinke, zadrži staru
+        ]);
+
+
+        // Povratni odgovor
+        return response()->json([
+            'message' => 'Sestra je uspešno ažurirana.',
+            'user' => $user,
+            'sestra' => $sestra,
+        ], 200);
     }
 
     /**
@@ -105,6 +112,7 @@ class SestraController extends Controller
     {
 
         $sestra = Sestra::findOrFail($id);
+        Gate::authorize('delete', $sestra);
         $user = User::findOrFail($sestra->user_id);
         $sestra->delete();
         $user->delete();
