+++ app/Http/Controllers/api/TerapijaController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api;
 use App\Models\Terapija;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\Resource\TerapijaResource;
 
 class TerapijaController extends Controller
@@ -14,6 +15,8 @@ class TerapijaController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny',Terapija::class);
+
         $query = Terapija::query();
         return TerapijaResource::collection($query->latest()->paginate());
     }
@@ -23,6 +26,8 @@ class TerapijaController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create',Terapija::class);
+
         $terapija = Terapija::create(
             $request->validate([
                 'naziv' => 'required|string|max:255'
@@ -38,6 +43,8 @@ class TerapijaController extends Controller
     public function show(string $id)
     {
         $terapija = Terapija::findOrFail($id);
+        Gate::authorize('view',$terapija);
+
         return new TerapijaResource($terapija);
     
     }
@@ -48,10 +55,14 @@ class TerapijaController extends Controller
     public function update(Request $request, string $id)
     {
         $terapija = Terapija::findOrFail($id);
+        Gate::authorize('update',$terapija);
+
         $terapija->update(
             $request->validate(
                 [
-                    'naziv' => 'required|string|max:255'
+                    'naziv' => 'required|string|max:255',
+                    'opis'=>'required|string|max:255',
+                    'trajanje'=>'required|integer'
                 ]
             )
         );
@@ -65,6 +76,8 @@ class TerapijaController extends Controller
     public function destroy(string $id)
     {
         $terapija = Terapija::findOrFail($id);
+        Gate::authorize('delete',$terapija);
+
         $terapija->delete();
         return response()->json('Successfully deleted');
     }
